requirements.txt
langchain-core>=0.1.0
langgraph>=0.0.1
langchain-openai>=0.0.1
pydantic>=2.0.0
python-dotenv>=1.0.0
langsmith>=0.0.1
httpx>=0.24.0
